完善视图功能

+++ app1/views/BorrowerViews/bookviews.py
@@ -3,7 +3,7 @@ from app1.form import BookForm
 from app1 import models
 
 
-def admin_book_list(request):
+def book_list(request):
     """
     图书列表
     :param request:
@@ -11,7 +11,9 @@ def admin_book_list(request):
     """
     # 图书列表
     books = models.Book.objects.all()
-    return render(request, 'AdminTemplates/AdBookTemplates/admin_list_book.html', {'books': books})
+    if request.session["user"] == "admin":
+        return render(request, 'AdminTemplates/AdBookTemplates/admin_list_book.html', {'books': books})
+    return render(request, 'CommonTemplates/BookTemplates/com_list_book.html', {'books': books})
 
 
 def admin_add_book(request):
@@ -20,6 +22,9 @@ def admin_add_book(request):
     :param request:
     :return:
     """
+    if request.session["user"] != "admin":
+        return render(request, 'CommonTemplates/BookTemplates/com_list_book.html')
+
     if request.method == 'GET':
         form = BookForm()
         return render(request, 'AdminTemplates/AdBookTemplates/admin_add_book.html', {'form': form})
@@ -29,7 +34,7 @@ def admin_add_book(request):
         book = form.save(commit=False)
         book.published_date = form.cleaned_data['ctime']
         book.save()
-        return redirect('admin_book_list')
+        return redirect('book_list')
     return render(request, 'AdminTemplates/AdBookTemplates/admin_add_book.html', {'form': form})
 
 
@@ -40,6 +45,9 @@ def admin_edit_book(request, book_id):
     :param book_id: 图书id
     :return:
     """
+    if request.session["user"] != "admin":
+        return render(request, 'CommonTemplates/BookTemplates/com_list_book.html')
+
     book = get_object_or_404(models.Book, pk=book_id)
     if request.method == 'GET':
         form = BookForm(instance=book, initial={'ctime': book.published_date})
@@ -50,7 +58,7 @@ def admin_edit_book(request, book_id):
             book = form.save(commit=False)
             book.published_date = form.cleaned_data['ctime']
             form.save()
-            return redirect('admin_book_list')
+            return redirect('book_list')
     return render(request, 'AdminTemplates/AdBookTemplates/admin_edit_book.html', {'form': form, 'book': book})
 
 
@@ -61,8 +69,14 @@ def admin_delete_book(request, book_id):
     :param book_id: 图书id
     :return:
     """
+    if request.session["user"] != "admin":
+        return render(request, 'CommonTemplates/BookTemplates/com_list_book.html')
+
     book = get_object_or_404(models.Book, pk=book_id)
     book.delete()
-    return redirect('admin_book_list')
+    return redirect('book_list')
+
+
+
